+++ user/src/main/java/qa/gov/custom/user/repository/RoleUserRepository.java
@@ -14,17 +14,17 @@ import java.util.List;
 
 @Repository
 @Transactional
-public interface RoleUserRepository extends JpaRepository<RoleUser, BigDecimal> {
+public interface RoleUserRepository extends JpaRepository<RoleUser, BigInteger> {
 
     @Modifying
     @Transactional
     @Query(value="insert into ROLE_USER(ROLE_ID,USER_ID) values(:roleId,:userId)",nativeQuery=true)
-    void insertUserRole(BigInteger userId , BigDecimal roleId);
+    void insertUserRole(BigInteger userId , BigInteger roleId);
 
     @Modifying
     @Transactional
     @Query(value="update  ROLE_USER SET ROLE_ID=:roleId where USER_ID=:userId",nativeQuery=true)
-    void updateUserRole(BigInteger userId , BigDecimal roleId);
+    void updateUserRole(BigInteger userId , BigInteger roleId);
 
     @Query(value="select ID,USER_ID,ROLE_ID from ROLE_USER  where USER_ID=:userId",nativeQuery=true)
     List<Object[]>  findRoleUserByUserId( BigInteger userId);
